+++ src/main/java/jscover/ConfigurationCommon.java
@@ -382,7 +382,7 @@ public class ConfigurationCommon extends Configuration {
     protected boolean isolateBrowser;
     protected final List<PatternMatcher> patternMatchers = new ArrayList<>();
     private boolean includeUnloadedJS;
-    protected Config.LanguageMode ECMAVersion = Config.LanguageMode.ECMASCRIPT8;
+    protected Config.LanguageMode ECMAVersion = Config.LanguageMode.ES_NEXT;
     protected boolean defaultSkip;
     protected IoUtils ioUtils = IoUtils.getInstance();
     protected Level logLevel = SEVERE;