+++ History.md
@@ -1,6 +1,7 @@
 2.0.6 / 2018-??-??
 ==================
-  * Upgrade closure-compiler v20180204 to v20180506 
+  * Upgrade closure-compiler v20180204 to v20180506
+  * Bump JS language in from ECMASCRIPT8 to ES_NEXT 
   * Internal: Mockito 2.18.0 to 2.18.3
 
 2.0.5 / 2018-04-11